[update] 調整紅利操作頁面

+++ src/EtheriT.Coker.Core/Models/Bonus.cs
@@ -4,11 +4,11 @@ namespace EtheriT.Coker.Core.Models
 {
     public class Bonus : FullAuditedEntity
     {
-        public double Amount { get; set; }
+        public int Amount { get; set; }
         public Guid UUID { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public string Note { get; set; }
-        public double Balance { get; set; }
+        public int Balance { get; set; }
     }
 }
